+++ HISTORY.md
@@ -1,3 +1,8 @@
+unreleased
+==========
+
+  * Fix `res.writeHead` patch missing return value
+
 1.0.1 / 2015-09-29
 ==================